update the osgi tests code for BROOKLYN-162 refactoring (replacing newly restored old files in original location with the updated files we made in the temporary location)
